test(streaming): isolate AssistantStream lifecycle regressions
